configuration: Fix warnings in legacy

+++ softwareComponents/configuration/legacy_include/legacy/configuration/Generators.h
@@ -10,18 +10,18 @@ std::optional<Configuration> executeIfValid(const Configuration& config, const A
 
 /**
  * \brief Generates all possible actions.
- * 
- * Fills \p res with all possible valid actions that contains up to \p bound rotations, connections 
+ *
+ * Fills \p res with all possible valid actions that contains up to \p bound rotations, connections
  * and disconnections in sum. Only rotations with angle + \p step and - \p step are allowed.
  */
 void generateActions(const Configuration& config, std::vector<Action>& res, unsigned step, unsigned bound=1);
 
 /**
  * \brief Generates all possible "simple" actions.
- * 
+ *
  * Simple action is an action that contains just a single rotation, connection and disconnection.
  * Only rotations with angle + \p step and - \p step are allowed.
- * 
+ *
  * Fills \p res with all possible valid simple actions
  */
 void generateSimpleActions(const Configuration& config, std::vector<Action>& res, unsigned step);
@@ -30,20 +30,20 @@ void generateSimpleOnlyRotActions(const Configuration& config, std::vector<Actio
 
 /**
  * \brief Generates all possible "bisimple" actions.
- * 
+ *
  * Bisimple action is an action that contains up to two rotations, or just a connection or a disconnection.
  * Only rotations with angle + \p step and - \p step are allowed.
- * 
+ *
  * Fills \p res with all possible valid bisimple actions
  */
 void generateBisimpleActions(const Configuration& config, std::vector<Action>& res, unsigned step);
 
 void generateBisimpleOnlyRotActions(const Configuration& config, std::vector<Action>& res, unsigned step);
 
-void generateParalyzedActions(const Configuration& config, std::vector<Action>& res, unsigned step, 
+void generateParalyzedActions(const Configuration& config, std::vector<Action>& res, unsigned step,
     const std::unordered_set<ID>& allowed_indices);
 
-void generateBiParalyzedOnlyRotAction(const Configuration& config, std::vector<Action>& res, unsigned step, 
+void generateBiParalyzedOnlyRotAction(const Configuration& config, std::vector<Action>& res, unsigned step,
     const std::unordered_set<ID>& allowed_indices);
 
 void generateSmartParalyzedOnlyRotActions(const Configuration& config, std::vector<Action>& res, unsigned step,
@@ -51,10 +51,10 @@ void generateSmartParalyzedOnlyRotActions(const Configuration& config, std::vect
 
 void generateRotations(const Configuration& config, std::vector<Action::Rotate>& res, unsigned step);
 
-void generateParalyzedRotations(const Configuration& config, std::vector<Action::Rotate>& res, unsigned step, 
+void generateParalyzedRotations(const Configuration& config, std::vector<Action::Rotate>& res, unsigned step,
     const std::unordered_set<ID>& allowed_indices);
 
-void generateMappedParalyzedRotations(const Configuration& config, std::unordered_map<ID, std::vector<Action::Rotate>>& res, unsigned step, 
+void generateMappedParalyzedRotations(const Configuration& config, std::unordered_map<ID, std::vector<Action::Rotate>>& res, unsigned step,
     const std::unordered_set<ID>& allowed_indices);
 
 void generateConnections(const Configuration& config, std::vector<Action::Reconnect>& res);
@@ -73,7 +73,7 @@ inline void generateReconnect(const Configuration& config, std::vector<Action::R
 }
 
 inline void generateParalyzedReconnect(const Configuration& config, std::vector<Action::Reconnect>& res,
-    const std::unordered_set<ID>& allowed_indices) 
+    const std::unordered_set<ID>& allowed_indices)
 {
     generateParalyzedConnections(config, res, allowed_indices);
     generateParalyzedDisconnections(config, res, allowed_indices);
@@ -81,19 +81,19 @@ inline void generateParalyzedReconnect(const Configuration& config, std::vector<
 
 /**
  * \brief Generates all possible configurations that \p config can become using one action.
- * 
- * Fills \p res with all possible configurations that \p config can change into using one action 
- * that contains up to \p bound rotations, connections and disconnections in sum. Only rotations 
+ *
+ * Fills \p res with all possible configurations that \p config can change into using one action
+ * that contains up to \p bound rotations, connections and disconnections in sum. Only rotations
  * with angle + \p step and - \p step are allowed.
- * 
+ *
  * If \p bound is equal to 1, it uses `generateSimpleActions` instead of `generateActions`.
  */
 void next(const Configuration& config, std::vector<Configuration>& res, unsigned step, unsigned bound=1);
 
 /**
  * \brief Generates all possible configurations that \p config can become using one simple action.
- * 
- * Fills \p res with all possible configurations that \p config can change into using one simple action 
+ *
+ * Fills \p res with all possible configurations that \p config can change into using one simple action
  * Only rotations with angle + \p step and - \p step are allowed.
  */
 void simpleNext(const Configuration& config, std::vector<Configuration>& res, unsigned step);
@@ -101,11 +101,11 @@ void simpleNext(const Configuration& config, std::vector<Configuration>& res, un
 void simpleOnlyRotNext(const Configuration& config, std::vector<Configuration>& res, unsigned step);
 
 /**
- * \brief Generates all possible configurations that \p config can become using one recconect action or 
+ * \brief Generates all possible configurations that \p config can become using one recconect action or
  * up to two simple rotations.
- * 
- * Fills \p res with all possible configurations that \p config can change into using one recconect 
- * action or up to two simple rotations. 
+ *
+ * Fills \p res with all possible configurations that \p config can change into using one recconect
+ * action or up to two simple rotations.
  * Only rotations with angle + \p step and - \p step are allowed.
  */
 void bisimpleNext(const Configuration& config, std::vector<Configuration>& res, unsigned step);
@@ -115,16 +115,16 @@ void bisimpleOnlyRotNext(const Configuration& config, std::vector<Configuration>
 /**
  * \brief Generates all possible configurations that \p config can become using one simple action that
  * can use only modules from \p allowed_indices.
- * 
+ *
  * Fills \p res with all possible configurations that \p config can change into using one simple action.
  * Only rotations of modules with id in \p allowed_indices are allowed. Only connections adn disconnections,
  * where at least one of the modules has its ID in \p allowed_indices are allowed.
  * Only rotations with angle + \p step and - \p step are allowed.
  */
-void paralyzedNext(const Configuration& config, std::vector<Configuration>& res, unsigned step, 
+void paralyzedNext(const Configuration& config, std::vector<Configuration>& res, unsigned step,
     const std::unordered_set<ID>& allowed_indices);
 
-void biParalyzedOnlyRotNext(const Configuration& config, std::vector<Configuration>& res, unsigned step, 
+void biParalyzedOnlyRotNext(const Configuration& config, std::vector<Configuration>& res, unsigned step,
     const std::unordered_set<ID>& allowed_indices);
 
 void smartBisimpleOnlyRotNext(const Configuration& config, std::vector<Configuration>& res, unsigned step);
@@ -133,8 +133,8 @@ void smartBisimpleParOnlyRotNext(const Configuration& config, std::vector<Config
     const std::unordered_set<ID>& allowed_indices);
 
 template<typename T>
-inline void getAllSubsetsRec(const std::vector<T>& set, std::vector<std::vector<T>>& res, 
-    const std::vector<T>& accum, unsigned index, unsigned count) 
+inline void getAllSubsetsRec(const std::vector<T>& set, std::vector<std::vector<T>>& res,
+    const std::vector<T>& accum, unsigned index, size_t count)
 {
     if ((accum.size() == count) || (index >= set.size())) {
         res.push_back(accum);
@@ -148,7 +148,7 @@ inline void getAllSubsetsRec(const std::vector<T>& set, std::vector<std::vector<
 
 using Rots = std::vector<Action::Rotate>;
 template<>
-inline void getAllSubsetsRec(const Rots& set, std::vector<Rots>& res, const Rots& accum, unsigned index, unsigned count) {
+inline void getAllSubsetsRec(const Rots& set, std::vector<Rots>& res, const Rots& accum, unsigned index, size_t count) {
     if ((accum.size() == count) || (index >= set.size())) {
         res.push_back(accum);
         return;
@@ -169,7 +169,7 @@ inline void getAllSubsetsRec(const Rots& set, std::vector<Rots>& res, const Rots
 }
 
 template<typename T>
-void getAllSubsets(const std::vector<T>& from, std::vector<std::vector<T>>& res, unsigned count) {
+void getAllSubsets(const std::vector<T>& from, std::vector<std::vector<T>>& res, size_t count) {
     getAllSubsetsRec<T>(from, res, {}, 0, count);
 }
